requirements.txt
# This file may be used to create an environment using:
# $ conda create --name xai --file requirements.txt python=3.9
numpy==1.22.3 
pandas==1.4.1
Pillow==9.0.1
matplotlib==3.5.1
seaborn==0.11.2
jupyterlab==3.3.1
ipywidgets==7.6.5
click==8.0.4
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
tensorboard==2.8.0
timm==0.5.4
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torch==1.11.0+cu113
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torchaudio==0.11.0+cu113
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torchmetrics==0.7.2
-f https://download.pytorch.org/whl/cu113/torch_stable.html
torchvision==0.12.0+cu113
pytorch-lightning==1.5.10
captum==0.5.0
tqdm==4.63.0
wandb==0.10.33
fire==0.4.0
minio==7.1.5
protobuf==3.19.4